+++ packages/core/schematics/migrations/output-migration/output-migration.ts
@@ -20,8 +20,8 @@ import {
   TsurgeFunnelMigration,
 } from '../../utils/tsurge';
 
-import {DtsMetadataReader} from '../../../../compiler-cli/src/ngtsc/metadata';
-import {TypeScriptReflectionHost} from '../../../../compiler-cli/src/ngtsc/reflection';
+import {DtsMetadataReader} from '@angular/compiler-cli/src/ngtsc/metadata';
+import {TypeScriptReflectionHost} from '@angular/compiler-cli/src/ngtsc/reflection';
 import {PartialEvaluator} from '@angular/compiler-cli/private/migrations';
 import {
   getUniqueIdForProperty,
@@ -148,13 +148,21 @@ export class OutputMigration extends TsurgeFunnelMigration<
                 outputFile,
               )
             ) {
-              filesWithOutputDeclarations.add(node.getSourceFile());
-              addOutputReplacement(
-                outputFieldReplacements,
-                outputDef.id,
-                outputFile,
-                calculateDeclarationReplacement(info, node, outputDef.aliasParam),
-              );
+              const aliasParam = outputDef.aliasParam;
+              const aliasOptionValue = aliasParam ? evaluator.evaluate(aliasParam) : undefined;
+
+              if (aliasOptionValue == undefined || typeof aliasOptionValue === 'string') {
+                filesWithOutputDeclarations.add(node.getSourceFile());
+                addOutputReplacement(
+                  outputFieldReplacements,
+                  outputDef.id,
+                  outputFile,
+                  calculateDeclarationReplacement(info, node, aliasOptionValue?.toString()),
+                );
+              } else {
+                problematicUsages[outputDef.id] = true;
+                problematicDeclarationCount++;
+              }
             }
           } else {
             problematicDeclarationCount++;