+++ packages/core/src/defer/instructions.ts
@@ -40,7 +40,6 @@ import {
   getTView,
   nextBindingIndex,
 } from '../render3/state';
-import {isPlatformBrowser} from '../render3/util/misc_utils';
 import {
   getConstant,
   getTNode,
@@ -90,6 +89,7 @@ import {onTimer, scheduleTimerTrigger} from './timer_scheduler';
 import {
   addDepsToRegistry,
   assertDeferredDependenciesLoaded,
+  assertIncrementalHydrationIsConfigured,
   getLDeferBlockDetails,
   getLoadingBlockAfter,
   getMinimumDurationForState,
@@ -99,7 +99,7 @@ import {
   setLDeferBlockDetails,
   setTDeferBlockDetails,
 } from './utils';
-import {DeferBlockRegistry} from './registry';
+import {DEFER_BLOCK_REGISTRY, DeferBlockRegistry} from './registry';
 import {incrementallyHydrateFromBlockName} from '../hydration/blocks';
 import {isIncrementalHydrationEnabled} from '../hydration/utils';
 
@@ -133,7 +133,7 @@ function shouldActivateHydrateTrigger(lView: LView, tNode: TNode): boolean {
   // LDeferBlockDetails as a flag. This would make subsequent lookups very cheap.
   return (
     isIncrementalHydrationEnabled(injector) &&
-    (!isPlatformBrowser(injector) || lDetails[SSR_UNIQUE_ID] !== null)
+    ((typeof ngServerMode !== 'undefined' && ngServerMode) || lDetails[SSR_UNIQUE_ID] !== null)
   );
 }
 
@@ -149,9 +149,10 @@ function shouldTriggerWhenOnClient(
   lDetails: LDeferBlockDetails,
   tDetails: TDeferBlockDetails,
 ): boolean {
-  if (!isPlatformBrowser(injector)) {
+  if (typeof ngServerMode !== 'undefined' && ngServerMode) {
     return false;
   }
+
   const isServerRendered =
     lDetails[SSR_BLOCK_STATE] && lDetails[SSR_BLOCK_STATE] === DeferBlockState.Complete;
   const hasHydrateTriggers = tDetails.hydrateTriggers && tDetails.hydrateTriggers.size > 0;
@@ -175,7 +176,12 @@ function shouldTriggerDeferBlock(
   if (config?.behavior === DeferBlockBehavior.Manual) {
     return false;
   }
-  return isPlatformBrowser(injector) || tDeferBlockDetails.hydrateTriggers !== null;
+
+  return (
+    typeof ngServerMode === 'undefined' ||
+    !ngServerMode ||
+    tDeferBlockDetails.hydrateTriggers !== null
+  );
 }
 
 function getHydrateTriggers(tView: TView, tNode: TNode): Map<DeferBlockTrigger, number | null> {
@@ -314,11 +320,11 @@ export function ɵɵdefer(
 
   let registry: DeferBlockRegistry | null = null;
   if (ssrUniqueId !== null) {
-    // TODO(incremental-hydration): explore how we can make
-    // `DeferBlockRegistry` tree-shakable for client-only cases.
-    registry = injector.get(DeferBlockRegistry);
+    ngDevMode && assertIncrementalHydrationIsConfigured(injector);
 
-    // Also store this defer block in the registry.
+    // Store this defer block in the registry, to have an access to
+    // internal data structures from hydration runtime code.
+    registry = injector.get(DEFER_BLOCK_REGISTRY);
     registry.add(ssrUniqueId, {lView, tNode, lContainer});
   }
 
@@ -412,7 +418,7 @@ export function ɵɵdeferHydrateWhen(rawValue: unknown) {
 
   if (bindingUpdated(lView, bindingIndex, rawValue)) {
     const injector = lView[INJECTOR]!;
-    if (!isPlatformBrowser(injector)) {
+    if (typeof ngServerMode !== 'undefined' && ngServerMode) {
       // We are on the server and SSR for defer blocks is enabled.
       triggerDeferBlock(lView, tNode);
     } else {
@@ -447,7 +453,7 @@ export function ɵɵdeferHydrateNever() {
     const hydrateTriggers = getHydrateTriggers(getTView(), tNode);
     hydrateTriggers.set(DeferBlockTrigger.Never, null);
 
-    if (!isPlatformBrowser(lView[INJECTOR]!)) {
+    if (typeof ngServerMode !== 'undefined' && ngServerMode) {
       // We are on the server and SSR for defer blocks is enabled.
       triggerDeferBlock(lView, tNode);
     }
@@ -481,7 +487,7 @@ export function ɵɵdeferHydrateOnIdle() {
     const hydrateTriggers = getHydrateTriggers(getTView(), tNode);
     hydrateTriggers.set(DeferBlockTrigger.Idle, null);
 
-    if (!isPlatformBrowser(lView[INJECTOR]!)) {
+    if (typeof ngServerMode !== 'undefined' && ngServerMode) {
       // We are on the server and SSR for defer blocks is enabled.
       triggerDeferBlock(lView, tNode);
     } else {
@@ -543,7 +549,7 @@ export function ɵɵdeferHydrateOnImmediate() {
     const hydrateTriggers = getHydrateTriggers(getTView(), tNode);
     hydrateTriggers.set(DeferBlockTrigger.Immediate, null);
 
-    if (!isPlatformBrowser(injector)) {
+    if (typeof ngServerMode !== 'undefined' && ngServerMode) {
       triggerDeferBlock(lView, tNode);
     } else {
       // TODO(incremental-hydration): see if we can resolve the circular dep issue
@@ -587,7 +593,7 @@ export function ɵɵdeferHydrateOnTimer(delay: number) {
     const hydrateTriggers = getHydrateTriggers(getTView(), tNode);
     hydrateTriggers.set(DeferBlockTrigger.Timer, delay);
 
-    if (!isPlatformBrowser(lView[INJECTOR]!)) {
+    if (typeof ngServerMode !== 'undefined' && ngServerMode) {
       // We are on the server and SSR for defer blocks is enabled.
       triggerDeferBlock(lView, tNode);
     } else {
@@ -659,7 +665,7 @@ export function ɵɵdeferHydrateOnHover() {
     const hydrateTriggers = getHydrateTriggers(getTView(), tNode);
     hydrateTriggers.set(DeferBlockTrigger.Hover, null);
 
-    if (!isPlatformBrowser(lView[INJECTOR]!)) {
+    if (typeof ngServerMode !== 'undefined' && ngServerMode) {
       // We are on the server and SSR for defer blocks is enabled.
       triggerDeferBlock(lView, tNode);
     }
@@ -731,7 +737,7 @@ export function ɵɵdeferHydrateOnInteraction() {
     const hydrateTriggers = getHydrateTriggers(getTView(), tNode);
     hydrateTriggers.set(DeferBlockTrigger.Interaction, null);
 
-    if (!isPlatformBrowser(lView[INJECTOR]!)) {
+    if (typeof ngServerMode !== 'undefined' && ngServerMode) {
       // We are on the server and SSR for defer blocks is enabled.
       triggerDeferBlock(lView, tNode);
     }
@@ -804,7 +810,7 @@ export function ɵɵdeferHydrateOnViewport() {
     hydrateTriggers.set(DeferBlockTrigger.Viewport, null);
     const injector = lView[INJECTOR]!;
 
-    if (!isPlatformBrowser(injector)) {
+    if (typeof ngServerMode !== 'undefined' && ngServerMode) {
       // We are on the server and SSR for defer blocks is enabled.
       triggerDeferBlock(lView, tNode);
     }
@@ -846,24 +852,26 @@ function scheduleDelayedPrefetching(
   scheduleFn: (callback: VoidFunction, injector: Injector) => VoidFunction,
   trigger: DeferBlockTrigger,
 ) {
+  if (typeof ngServerMode !== 'undefined' && ngServerMode) {
+    return;
+  }
+
   const lView = getLView();
   const injector = lView[INJECTOR]!;
 
   // Only trigger the scheduled trigger on the browser
   // since we don't want to delay the server response.
-  if (isPlatformBrowser(injector)) {
-    const tNode = getCurrentTNode()!;
-    const tView = lView[TVIEW];
-    const tDetails = getTDeferBlockDetails(tView, tNode);
-    const prefetchTriggers = getPrefetchTriggers(tDetails);
-    prefetchTriggers.add(trigger);
+  const tNode = getCurrentTNode()!;
+  const tView = lView[TVIEW];
+  const tDetails = getTDeferBlockDetails(tView, tNode);
+  const prefetchTriggers = getPrefetchTriggers(tDetails);
+  prefetchTriggers.add(trigger);
 
-    if (tDetails.loadingState === DeferDependenciesLoadingState.NOT_STARTED) {
-      const lDetails = getLDeferBlockDetails(lView, tNode);
-      const prefetch = () => triggerPrefetching(tDetails, lView, tNode);
-      const cleanupFn = scheduleFn(prefetch, injector);
-      storeTriggerCleanupFn(TriggerType.Prefetch, lDetails, cleanupFn);
-    }
+  if (tDetails.loadingState === DeferDependenciesLoadingState.NOT_STARTED) {
+    const lDetails = getLDeferBlockDetails(lView, tNode);
+    const prefetch = () => triggerPrefetching(tDetails, lView, tNode);
+    const cleanupFn = scheduleFn(prefetch, injector);
+    storeTriggerCleanupFn(TriggerType.Prefetch, lDetails, cleanupFn);
   }
 }
 
@@ -875,22 +883,24 @@ export function scheduleDelayedHydrating(
   lView: LView,
   tNode: TNode,
 ) {
+  if (typeof ngServerMode !== 'undefined' && ngServerMode) {
+    return;
+  }
+
   // Only trigger the scheduled trigger on the browser
   // since we don't want to delay the server response.
   const injector = lView[INJECTOR]!;
-  if (isPlatformBrowser(injector)) {
-    const lDetails = getLDeferBlockDetails(lView, tNode);
-    const cleanupFn = scheduleFn(
-      () =>
-        incrementallyHydrateFromBlockName(
-          injector,
-          lDetails[SSR_UNIQUE_ID]!,
-          (deferBlock: DeferBlock) => triggerAndWaitForCompletion(deferBlock),
-        ),
-      injector,
-    );
-    storeTriggerCleanupFn(TriggerType.Hydrate, lDetails, cleanupFn);
-  }
+  const lDetails = getLDeferBlockDetails(lView, tNode);
+  const cleanupFn = scheduleFn(
+    () =>
+      incrementallyHydrateFromBlockName(
+        injector,
+        lDetails[SSR_UNIQUE_ID]!,
+        (deferBlock: DeferBlock) => triggerAndWaitForCompletion(deferBlock),
+      ),
+    injector,
+  );
+  storeTriggerCleanupFn(TriggerType.Hydrate, lDetails, cleanupFn);
 }
 
 /**
@@ -936,12 +946,11 @@ export function renderDeferBlockState(
     isValidStateChange(currentState, newState) &&
     isValidStateChange(lDetails[NEXT_DEFER_BLOCK_STATE] ?? -1, newState)
   ) {
-    const injector = hostLView[INJECTOR]!;
     const tDetails = getTDeferBlockDetails(hostTView, tNode);
     // Skips scheduling on the server since it can delay the server response.
     const needsScheduling =
       !skipTimerScheduling &&
-      isPlatformBrowser(injector) &&
+      (typeof ngServerMode === 'undefined' || !ngServerMode) &&
       (getLoadingBlockAfter(tDetails) !== null ||
         getMinimumDurationForState(tDetails, DeferBlockState.Loading) !== null ||
         getMinimumDurationForState(tDetails, DeferBlockState.Placeholder));
@@ -1436,7 +1445,7 @@ export function triggerDeferBlock(lView: LView, tNode: TNode) {
 
   let registry: DeferBlockRegistry | null = null;
   if (isIncrementalHydrationEnabled(injector)) {
-    registry = injector.get(DeferBlockRegistry);
+    registry = injector.get(DEFER_BLOCK_REGISTRY);
   }
   // Defer block is triggered, cleanup all registered trigger functions.
   invokeAllTriggerCleanupFns(lDetails, registry);